Move a few bits out of utils
